refactor: move usePlayingTrackProgress
