+++ apps/rtk-query/src/pages/TracksPage/TracksPage.tsx
@@ -11,8 +11,8 @@ import {
 import { TrackActions } from '@/features/tracks/ui/TrackActions/TrackActions'
 import { TrackRow } from '@/features/tracks/ui/TrackRow/TrackRow'
 import { useCurrentTrack, usePlaybackState, usePlayerControls } from '@/player'
+import { usePlayingTrackProgress } from '@/player/playerHooks.ts'
 import { convertApiTracksToPlayerTracks, convertApiTrackToPlayerTrack } from '@/player/utils.ts'
-import { usePlayingTrackProgress } from '@/player/utils/getPlayingTrackProgress.ts'
 import noCoverPlaceholder from '@/shared/assets/images/no-cover-placeholder.avif'
 import { Typography } from '@/shared/components'
 import { Spinner } from '@/shared/components/Spinner/Spinner.tsx'
